requirements.txt
alabaster==0.7.12
alembic==1.0.7
amqp==2.4.1
aniso8601==4.1.0
apispec==3.2.0
attrs==19.3.0
Babel==2.6.0
backcall==0.1.0
bcrypt==3.1.7
billiard==3.5.0.5
blinker==1.4
boto3==1.9.238
botocore==1.12.238
celery==4.1.1
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
Click==7.0
cryptography==3.4.7
curlify==2.2.1
decorator==4.3.2
defusedxml==0.6.0
dnspython==1.16.0
docutils==0.15.2
elasticsearch==7.0.0
et-xmlfile==1.0.1
eventlet==0.25.1
Flask==1.0.2
Flask-Admin==1.5.3
flask-apispec==0.8.5
Flask-BabelEx==0.9.3
Flask-Bcrypt==0.7.1
flask-blueprint==1.2.7
Flask-Celery-Helper==1.1.0
Flask-Cors==3.0.7
Flask-Elasticsearch==0.2.5
Flask-Excel==0.0.7
Flask-JWT-Extended==3.23.0
Flask-Limiter==1.0.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Mail-SendGrid==0.3
flask-marshmallow==0.10.1
Flask-Migrate==2.4.0
Flask-Principal==0.4.0
Flask-Redis==0.3.0
Flask-RESTful==0.3.7
Flask-Script==2.0.6
Flask-Security==3.0.0
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
GeoAlchemy2==0.6.3
geographiclib==1.50
geopy==1.20.0
gevent==20.9.0
greenlet==0.4.17
gunicorn==19.9.0
httplib2==0.18.1
idna==2.8
imagesize==1.2.0
importlib-metadata==1.3.0
ipython==7.3.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.13.3
Jinja2==2.10
jmespath==0.9.4
kombu==4.3.0
limits==1.3
lml==0.0.9
Mako==1.0.7
MarkupPy==1.14
MarkupSafe==1.1.1
marshmallow==2.20.5
marshmallow-sqlalchemy==0.21.0
monotonic==1.5
more-itertools==8.0.2
newrelic==5.22.1.152
numpy==1.19.1
oauth2==1.9.0.post1
odfpy==1.4.0
openpyxl==3.0.0
packaging==19.2
pandas==1.0.5
parso==0.3.4
passlib==1.7.1
pexpect==4.6.0
phonenumbers==8.12.31
pickleshare==0.7.5
pika==0.13.0
pluggy==0.13.1
prompt-toolkit==2.0.9
psycopg2-binary==2.8.3
ptyprocess==0.6.0
py==1.8.0
PyAMQP==0.0.8.11
pycparser==2.19
pyexcel==0.5.12
pyexcel-io==0.5.14
pyexcel-webio==0.1.4
pyexcel-xls==0.5.8
pyfcm==1.4.7
Pygments==2.3.1
PyJWT==1.7.1
PyMySQL==0.9.3
pyotp==2.3.0
pyparsing==2.4.5
pytest==5.3.2
python-dateutil==2.8.0
python-dotenv==0.15.0
python-editor==1.0.4
python-http-client==3.2.1
pytz==2018.9
PyYAML==5.1.2
raven==6.10.0
razorpay==1.2.0
redis==3.2.1
requests==2.22.0
s3transfer==0.2.1
sendgrid==6.1.0
sentry-sdk==0.13.5
Shapely==1.6.4.post2
simplejson==3.16.0
six==1.12.0
snowballstemmer==2.1.0
speaklater==1.3
Sphinx==4.2.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
SQLAlchemy==1.3.11
tablib==0.14.0
texttable==1.6.1
traitlets==4.3.2
twilio==6.33.1
urllib3==1.25.2
vine==1.2.0
wcwidth==0.1.7
webargs==5.5.3
Werkzeug==0.16.0
WTForms==2.2.1
xlrd==1.2.0
xlwt==1.3.0
zipp==0.6.0
zope.event==4.5.0
zope.interface==5.1.2